+++ NEWS.md
@@ -1,6 +1,8 @@
 # nycgeo 0.1.0.9000
 
 * Updated `nycgeo::nta_acs_data` to include properly aggregated data from `nycgeo::tract_acs_data`. Also, removed NTAs without population (e.g. parks, cemeteries, airports) from `nycgeo::tract_acs_data`.
+* Fixed bug in `nyc_boundaries()` when filtering by cd.
+* Added tests.
 * Added a `NEWS.md` file to track changes to the package.
 * Added a package logo.
 * Updated ACS datasets documentation.